% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meuc-data.R
\docType{data}
\name{dat1}
\alias{dat1}
\title{Dataset with Simulated (Y, C) Values for Examples in logreg_xerrors1}
\source{
Simulated data in R.
}
\description{
Includes 200 observations. The Xtilde values are stored as a separate list 
called \code{\link{dat1_xtilde}}. The data were generated with a true log-OR 
of 0.2 for X and Y, adjusted for C. The Xtilde measurements are subject to 
additive normal measurement error.
}
