% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meuc-data.R
\docType{data}
\name{dat_gdfa}
\alias{dat_gdfa}
\title{Dataset for Examples in gdfa}
\source{
Simulated data in R.
}
\description{
List of length 2 containing data frame with (Y, X, Xtilde, C) values for
250 subjects and list of Xtilde values where 10 subjects have replicates.
Xtilde values are affected by multiplicative lognormal measurement error.
}
