% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meuc-data.R
\docType{data}
\name{dat_logreg_xerrors3}
\alias{dat_logreg_xerrors3}
\title{Dataset for Examples in logreg_xerrors3}
\source{
Simulated data in R.
}
\description{
List of length 2 containing data frame with (Y, X, Xtilde, C) values for 500
subjects and list of Xtilde values where 25 subjects have replicates. Xtilde
values are affected by multiplicative lognormal measurement error.
}
