% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdfa_constant.R
\name{gdfa_constant}
\alias{gdfa_constant}
\title{Gamma Discriminant Function Approach for Estimating Odds Ratio with Exposure
Potentially Subject to Multiplicative Lognormal Measurement Error (Constant
Odds Ratio Version)}
\usage{
gdfa_constant(y, xtilde, c = NULL, merror = FALSE, integrate_tol = 1e-08,
  integrate_tol_hessian = integrate_tol, estimate_var = TRUE,
  fix_posdef = TRUE, ...)
}
\arguments{
\item{y}{Numeric vector of Y values.}

\item{xtilde}{Numeric vector (or list of numeric vectors, if there are
replicates) of Xtilde values.}

\item{c}{Numeric matrix with \strong{C} values (if any), with one row for
each subject. Can be a vector if there is only 1 covariate.}

\item{merror}{Logical value for whether there is measurement error.}

\item{integrate_tol}{Numeric value specifying the \code{tol} input to
\code{\link[cubature]{hcubature}}.}

\item{integrate_tol_hessian}{Same as \code{integrate_tol}, but for use when
estimating the Hessian matrix only. Sometimes more precise integration
(i.e. smaller tolerance) helps prevent cases where the inverse Hessian is not
positive definite.}

\item{estimate_var}{Logical value for whether to return variance-covariance
matrix for parameter estimates.}

\item{fix_posdef}{Logical value for whether to repeatedly reduce
\code{integrate_tol_hessian} by factor of 5 and re-estimate Hessian to try
to avoid non-positive definite variance-covariance matrix.}

\item{...}{Additional arguments to pass to \code{\link[stats]{nlminb}}.}
}
\value{
List containing:
\enumerate{
\item Numeric vector of parameter estimates.
\item Variance-covariance matrix.
\item Returned \code{\link[stats]{nlminb}} object from maximizing the
log-likelihood function.
\item Akaike information criterion (AIC).
}
}
\description{
See \code{\link{gdfa}}.
}
\references{
Whitcomb, B.W., Perkins, N.J., Zhang, Z., Ye, A., and Lyles, R. H. (2012)
"Assessment of skewed exposure in case-control studies with pooling."
\emph{Stat. Med.} \strong{31}: 2461--2472.
}
