% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_logistic_loglinear.R
\name{ml_logistic_loglinear}
\alias{ml_logistic_loglinear}
\title{Maximum Likelihood with Logistic Regression Disease Model and
Log-Transformed Linear Regression Measurement Error Model}
\usage{
ml_logistic_loglinear(all_data = NULL, main = NULL, internal = NULL,
  external = NULL, y_var, z_var, d_vars = NULL, c_vars = NULL,
  b_vars = NULL, tdm_covariates = NULL, mem_covariates = NULL,
  approx_integral = TRUE, integrate_tol = 1e-08,
  integrate_tol_hessian = integrate_tol, estimate_var = TRUE, ...)
}
\arguments{
\item{all_data}{Data frame with data for main study and validation study.}

\item{main}{Data frame with data for the main study.}

\item{internal}{Data frame with data for internal validation study.}

\item{external}{Data frame with data for the external validation study.}

\item{y_var}{Character string specifying name of Y variable.}

\item{z_var}{Character string specifying name of Z variable.}

\item{d_vars}{Character string specifying name of \strong{D} variables.}

\item{c_vars}{Character vector specifying names of \strong{C} variables.}

\item{b_vars}{Character vector specifying names of variables in true disease
model but not in measurement error model.}

\item{tdm_covariates}{Character vector specifying variables in true disease
model. The Z variable is automatically included whether you include it in
\code{tdm_covariates} or not.}

\item{mem_covariates}{Character vector specifying variables in measurement
error model.}

\item{approx_integral}{Logical value for whether to use the probit
approximation for the logistic-normal integral, to avoid numerically
integrating \code{Z}'s out of the likelihood function.}

\item{integrate_tol}{Numeric value specifying \code{tol} input to
\code{\link[cubature]{hcubature}} for numerical integration.}

\item{integrate_tol_hessian}{Same as \code{integrate_tol}, but for use when
estimating the Hessian matrix only. Sometimes using a smaller value than for
likelihood maximization helps prevent cases where the inverse Hessian is not
positive definite.}

\item{estimate_var}{Logical value for whether to return variance-covariance
matrix for parameter estimates.}

\item{...}{Additional arguments to pass to \code{\link[stats]{nlminb}}.}
}
\value{
List containing:
\enumerate{
\item Numeric vector of parameter estimates.
\item Variance-covariance matrix (if \code{estimate_var = TRUE}).
\item Returned \code{\link[stats]{nlminb}} object from maximizing the
log-likelihood function.
\item Akaike information criterion (AIC).
}
}
\description{
Calculates maximum likelihood estimates for measurement error/unmeasured
confounding scenario where true disease model is logistic regression and
measurement error model is log-transformed linear regression.
}
\details{
The true disease model is:

logit[P(Y = 1)] = beta_0 + beta_z Z + \strong{beta_c}^T \strong{C} +
\strong{beta_b}^T \strong{B}

The measurement error model is:

ln(Z) = alpha_0 + \strong{alpha_d}^T \strong{D} + \strong{alpha_c}^T
\strong{C} + d, d ~ N(0, sigsq_d)

There should be main study data with (Y, \strong{D}, \strong{C}, \strong{B})
as well as internal validation data with
(Y, Z, \strong{D}, \strong{C}, \strong{B}) and/or external validation data
with (Z, \strong{D}, \strong{C}).
}
