% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndfa.R
\name{ndfa}
\alias{ndfa}
\title{Normal Discriminant Function Approach for Estimating Odds Ratio with Exposure
Potentially Subject to Additive Normal Measurement Error}
\usage{
ndfa(y, xtilde, c = NULL, constant_or = TRUE, merror = TRUE, ...)
}
\arguments{
\item{y}{Numeric vector of Y values.}

\item{xtilde}{List of numeric vectors with Xtilde values.}

\item{c}{Numeric matrix with \strong{C} values (if any), with one row for
each subject. Can be a vector if there is only 1 covariate.}

\item{constant_or}{Logical value for whether to assume a constant odds ratio
for X, which means that sigsq_1 = sigsq_0. If \code{NULL}, model is fit with
and without this assumption, and likelihood ratio test is performed to test
it.}

\item{merror}{Logical value for whether there is measurement error.}

\item{...}{Additional arguments to pass to \code{\link[stats]{nlminb}}.}
}
\value{
List containing:
\enumerate{
\item Numeric vector of parameter estimates.
\item Variance-covariance matrix.
\item Returned \code{\link[stats]{nlminb}} object from maximizing the
log-likelihood function.
\item Akaike information criterion (AIC).
}

If \code{constant_or = NULL}, two such lists are returned (one under a
constant odds ratio assumption and one not), along with a likelihood ratio
test for \code{H0: sigsq_1 = sigsq_0}, which is equivalent to
\code{H0: odds ratio is constant}.
}
\description{
Assumes exposure given covariates and outcome is a normal-errors linear
regression. Exposure measurements can be assumed precise or subject to
additive normal measurement error. Some replicates are required for
identifiability. Parameters are estimated using maximum likelihood.
}
\examples{
# Load data frame with (Y, X, Xtilde, C) values for 1,000 subjects and list
# of Xtilde values where 25 subjects have replicates. Xtilde values are
# affected by measurement error. True log-OR = 0.5, sigsq = 1, sigsq_m = 0.5.
data(dat_ndfa)
dat <- dat_ndfa$dat
reps <- dat_ndfa$reps

# Estimate log-OR for X and Y adjusted for C using true X values
# (unobservable truth).
fit.unobservable <- ndfa(
  y = dat$y,
  xtilde = dat$x,
  c = dat$c,
  merror = FALSE
)
fit.unobservable$estimates

# Estimate log-OR for X and Y adjusted for C using observed Xtilde values,
# ignoring measurement error.
fit.naive <- ndfa(
  y = dat$y,
  xtilde = dat$xtilde,
  c = dat$c,
  merror = FALSE
)
fit.naive$estimates

# Repeat, but accounting for measurement error.
fit.corrected <- ndfa(
  y = dat$y,
  xtilde = reps,
  c = dat$c,
  merror = TRUE
)
fit.corrected$estimates

# Same as previous, but allowing for non-constant odds ratio.
fit.nonconstant <- ndfa(
  y = dat$y,
  xtilde = reps,
  c = dat$c,
  constant_or = FALSE,
  merror = TRUE
)
fit.nonconstant$estimates

# Perform likelihood ratio test for H0: odds ratio is constant.
lrt <- ndfa(
  y = dat$y,
  xtilde = reps,
  c = dat$c,
  constant_or = NULL,
  merror = TRUE
)


}
\references{
Lyles, R.H., Guo, Y., and Hill, A.N. (2009) "A fresh look at the discriminant
function approach for estimating crude or adjusted odds ratios."
\emph{Am. Stat} \strong{63}(4): 320--327.
}
