% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndfa_constant.R
\name{ndfa_constant}
\alias{ndfa_constant}
\title{Normal Discriminant Function Approach for Estimating Odds Ratio with Exposure
Potentially Subject to Additive Normal Measurement Error (Constant Odds Ratio
Version)}
\usage{
ndfa_constant(y, xtilde, c = NULL, merror = TRUE, ...)
}
\arguments{
\item{y}{Numeric vector of Y values.}

\item{xtilde}{List of numeric vectors with Xtilde values.}

\item{c}{Numeric matrix with \strong{C} values (if any), with one row for
each subject. Can be a vector if there is only 1 covariate.}

\item{merror}{Logical value for whether there is measurement error.}

\item{...}{Additional arguments to pass to \code{\link[stats]{nlminb}}.}
}
\value{
List containing:
\enumerate{
 \item Numeric vector of parameter estimates.
 \item Object returned by \code{\link[stats]{nlminb}} containing information
 about likelihood maximization.
 \item Akaike information criterion (AIC).
}
}
\description{
See \code{\link{ndfa}}.
}
\references{
Lyles, R.H., Guo, Y., and Hill, A.N. (2009) "A fresh look at the discriminant
function approach for estimating crude or adjusted odds ratios."
\emph{Am. Stat} \strong{63}(4): 320--327.
}
