% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psc.R, R/psc_linreg.R
\name{psc}
\alias{psc}
\alias{psc}
\title{Propensity Score Calibration for Linear Regression}
\usage{
psc(all_data = NULL, main = NULL, internal = NULL, external = NULL,
  y_var, x_var, gs_vars, ep_vars, tdm_family = "gaussian", surrogacy = TRUE,
  weighted = FALSE, ep_refit = FALSE)

psc(all_data = NULL, main = NULL, internal = NULL, external = NULL,
  y_var, x_var, gs_vars, ep_vars, tdm_family = "gaussian", surrogacy = TRUE,
  weighted = FALSE, ep_refit = FALSE)
}
\description{
Implements propensity score calibration as described by Sturmer et al.
(\strong{Am. J. Epidemiol.} 2005). Requires validation data.

Implements propensity score calibration as described by Sturmer et al.
(\strong{Am. J. Epidemiol.} 2005). Requires validation data.
}
\details{
The disease model is a GLM:

g[E(Y)] = beta_0 + beta_x X + beta_g G

where G = P(X|\strong{C},\strong{Z}), with \strong{C} but not \strong{Z}
available in the main study.

In a validation study with (X, \strong{C}, \strong{Z}), logistic regression
is used to obtain fitted probabilities for G as well as an error-prone
version G* = P(X|\strong{C}). A linear model is fitted to map from G* to
E(G|G*). Finally, in the main study, G*'s are calculated, then E(G|G*),
and the disease model is fit for Y vs. (X, E(G|G*)).

The disease model is:

E(Y) = beta_0 + beta_x X + beta_g G

where G = P(X|\strong{C},\strong{Z}), with \strong{C} but not \strong{Z}
available in the main study.

In a validation dataset with (X, \strong{C}, \strong{Z}), logistic regression
is used to obtain fitted probabilities for G as well as an error-prone
propensity score G* = P(X|\strong{C}). A linear model is fitted to map from
G* to E(G|G*). Finally, in the main study, G*'s are calculated, then E(G|G*),
and the disease model is fit for Y vs. (X, E(G|G*)).
}
\references{
Sturmer, T., Schneeweiss, S., Avorn, J. and Glynn, R.J. (2005) "Adjusting
effect estimates for unmeasured confounding with validation data using
propensity score calibration." \emph{Am. J. Epidemiol.} \strong{162}(3):
279-289.

Sturmer, T., Schneeweiss, S., Avorn, J. and Glynn, R.J. (2005) "Adjusting
effect estimates for unmeasured confounding with validation data using
propensity score calibration." \emph{Am. J. Epidemiol.} \strong{162}(3):
279-289.
}
