% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psc_algebraic.R
\name{psc_algebraic}
\alias{psc_algebraic}
\title{Propensity Score Calibration (Algebraic Method)}
\usage{
psc_algebraic(all_data = NULL, main = NULL, internal = NULL,
  external = NULL, y_var, x_var, gs_vars, ep_vars, tdm_family = "gaussian",
  ep_data = "validation", delta_var = TRUE, boot_var = FALSE,
  boots = 100, alpha = 0.05)
}
\arguments{
\item{all_data}{Data frame with data for main study and validation study.}

\item{main}{Data frame with data for the main study.}

\item{internal}{Data frame with data for internal validation study.}

\item{external}{Data frame with data for the external validation study.}

\item{y_var}{Character string specifying name of Y variable.}

\item{x_var}{Character string specifying name of X variable.}

\item{gs_vars}{Character vector specifying names of variables for gold
standard propensity score.}

\item{ep_vars}{Character vector specifying names of variables for error-prone
propensity score.}

\item{tdm_family}{Character string specifying family of true disease model
(see \code{\link[stats]{glm}}).}

\item{ep_data}{Character string controlling what data is used to fit the
error-prone propensity score model. Choices are \code{"validation"} for
validation study data, \code{"all"} for main study and validation study data,
and \code{"separate"} for validation data for first step and main study data
for second step.}

\item{delta_var}{Logical value for whether to calculate a Delta method
variance-covariance matrix. May not be justified theoretically because
propensity scores are estimated, but also may perform better than bootstrap
in certain scenarios, e.g. if bootstrap is prone to extreme estimates.}

\item{boot_var}{Logical value for whether to calculate a bootstrap
variance-covariance matrix.}

\item{boots}{Numeric value specifying number of bootstrap samples to use.}

\item{alpha}{Significance level for percentile bootstrap confidence interval.}
}
\description{
Implements propensity score calibration as described by Sturmer et al.
(\emph{Am. J. Epidemiol.} 2005). Requires validation data.
}
\details{
The true disease model is a GLM:

g[E(Y)] = beta_0 + beta_x X + beta_g G

where G = P(X|\strong{Z},\strong{C}), with \strong{C} but not \strong{Z}
available in the main study. Logistic regression models are used to model
P(X = 1|\strong{Z},\strong{C}) and P(X = 1|\strong{C}):

logit[P(X = 1)] = gamma_0 + \strong{gamma_z}^T \strong{Z} +
\strong{gamma_c}^T \strong{C}
logit[P(X = 1)] = gamma_0^* + \strong{gamma_c}^*T \strong{C}

A linear model is used to map fitted error-prone propensity scores Gstar
(from the X|\strong{C} logistic regression) to the expected gold standard
propensity score G (from the X|(\strong{C},\strong{Z}) logistic regression):

E(G) = lambda_0 + lambda_x X + lambda_g Gstar

There should be main study data with
(Y, X, \strong{C}) as well as external validation data with
(X, \strong{C}, \strong{Z}).
}
\references{
Sturmer, T., Schneeweiss, S., Avorn, J. and Glynn, R.J. (2005) "Adjusting
effect estimates for unmeasured confounding with validation data using
propensity score calibration." \emph{Am. J. Epidemiol.} \strong{162}(3):
279-289.
}
