% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_algebraic.R
\name{rc_algebraic}
\alias{rc_algebraic}
\title{Regression Calibration (Algebraic Method)}
\usage{
rc_algebraic(all_data = NULL, main = NULL, internal = NULL,
  external = NULL, y_var, z_var, d_var = NULL, c_vars = NULL,
  b_vars = NULL, tdm_covariates = NULL, tdm_family = "gaussian",
  mem_covariates = NULL, beta_0_formula = 1, delta_var = FALSE,
  boot_var = FALSE, boots = 100)
}
\arguments{
\item{all_data}{Data frame with data for main study and validation study.}

\item{main}{Data frame with data for the main study.}

\item{internal}{Data frame with data for internal validation study.}

\item{external}{Data frame with data for the external validation study.}

\item{y_var}{Character string specifying name of Y variable.}

\item{z_var}{Character string specifying name of Z variable.}

\item{d_var}{Character string specifying name of D variable.}

\item{c_vars}{Character vector specifying names of C variables.}

\item{b_vars}{Character vector specifying names of variables in true disease
model but not in measurement error model.}

\item{tdm_covariates}{Character vector specifying variables in true disease
model. The Z variable is automatically included whether you include it in
\code{tdm_covariates} or not.}

\item{tdm_family}{Character string specifying family of true disease model
(see \code{\link[stats]{glm}}).}

\item{mem_covariates}{Character vector specifying variables in measurement
error model.}

\item{beta_0_formula}{If \code{1}, formula for true disease model intercept
is:

beta_0.hat = betastar_0.hat - alpha_0.hat beta_Z.hat

If \code{2}, formula is:

beta_0.hat = betastar_0.hat - alpha_0.hat beta_Z.hat - 1/2 beta_Z.hat^2
sigma_delta^2

Formula 1 yields the same beta_0.hat as the "conditional expectation" view of
regression calibration (see \code{\link{rc_cond_exp}}). When the disease
model is logistic regression and the measurement error model is linear
regression, formula 1 is appropriate if Y is rare and Z|(D,\strong{C}) is
normal, and formula 2 is appropriate if beta_Z^2 sigma_delta^2 is small
(Kuha, Stat. Med. 1994). If neither criteria is met, regression calibration
may be unreliable.}

\item{delta_var}{Logical value for whether to calculate a Delta method
variance-covariance matrix.}

\item{boot_var}{Logical value for whether to calculate a bootstrap
variance-covariance matrix.}

\item{boots}{Numeric value specifying number of bootstrap samples to use.}
}
\value{
If no variance estimates are requested, a named numeric vector of parameter
estimates. If one or more variance estimates are requested, a list that also
contains a variance-covariance matrix for each variance estimator.
}
\description{
Implements the "algebraic" version of regression calibration as described in
Rosner et al. (\emph{Stat. Med.} 1989). For the "conditional expectation"
version, see \code{\link{rc_cond_exp}}.
}
\details{
The true disease model is a GLM:

g[E(Y)] = beta_0 + beta_z Z + \strong{beta_c}^T \strong{C} +
\strong{beta_b}^T \strong{B}

The measurement error model is:

E(Z) = alpha_d D + \strong{alpha_c}^T \strong{C}

And the naive disease model is:

g[E(Y)] = beta*_0 + beta*_Z D + \strong{beta*_C}^T \strong{C} +
\strong{beta*_B}^T \strong{B}

The procedure involves fitting the naive disease model using main study
data, fitting the measurement error model using validation data, and
solving a system of equations to get the regression calibration estimates.
}
\references{
Kuha, J. (1994) "Corrections for exposure measurement error in logistic
regression models with an application to nutritional data." \emph{Stat. Med.}
\strong{13}(11): 1135-1148.

Lyles, R.H. and Kupper, L.L. (2012) "Approximate and pseudo-likelihood
analysis for logistic regression using external validation data to model log
exposure." \emph{J. Agric. Biol. Environ. Stat.} \strong{18}(1): 22-38.

Rosner, B., Willett, W. and Spiegelman, D. (1989) "Correction of logistic
regression relative risk estimates and confidence intervals for systematic
within-person measurement error." \emph{Stat. Med.} \strong{8}(9): 1051-69.

Spiegelman, D., Carroll, R.J. and Kipnis, V. (2001) "Efficient regression
calibration for logistic regression in main study/internal validation study
designs with an imperfect reference instrument." \emph{Stat. Med.}
\strong{20}(1): 139-160.
}
