% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_greenland.R
\name{rc_greenland}
\alias{rc_greenland}
\title{Regression Calibration with Internal Validation Data and One Surrogate}
\usage{
rc_greenland(all_data = NULL, main = NULL, internal = NULL, y_var, z_var,
  d_var, c_vars = NULL, b_vars = NULL, tdm_family = "gaussian")
}
\arguments{
\item{all_data}{Data frame with data for main study and validation study.}

\item{main}{Data frame with data for the main study.}

\item{internal}{Data frame with data for internal validation study.}

\item{y_var}{Character string specifying name of Y variable.}

\item{z_var}{Character string specifying name of Z variable.}

\item{d_var}{Character string specifying name of D variable.}

\item{c_vars}{Character vector specifying names of \strong{C} variables.}

\item{b_vars}{Character vector specifying names of variables in true disease
model but not in measurement error model.}

\item{tdm_family}{Character string specifying family of true disease model
(see \code{\link[stats]{glm}}).}
}
\value{
List containing parameter estimates and variance estimates.
}
\description{
Implements an efficient version of regression calibration for
main study/internal validation designs with one surrogate, as described by
Spiegelman et al. (\emph{Stat. Med.} 2001). Uses ideas from Greenland
(\emph{Stat. Med.} 1988).
}
\details{
The true disease model is a GLM:

g[E(Y)] = beta_0 + beta_z Z + beta_c^T C + beta_b^T B

The measurement error model is:

E(Z) = alpha_d D + alpha_c^T C

And the naive disease model is:

g[E(Y)] = beta*_0 + beta*_Z D + \strong{beta*_C}^T \strong{C} +
\strong{beta*_B}^T \strong{B}

The procedure involves obtaining two sets of \strong{beta} estimates: one
by fitting the true disease model with internal validation data, and the
other by treating the validation study as external and doing a main
study/external validation study correction. The two estimates are weighted by
the inveres of their estimated variances.
}
\references{
Greenland, S. (1988) "Variance estimation for epidemiologic effect estimates
under misclassification." \emph{Stat. Med.} \strong{7}: 745-757.

Spiegelman, D., Carroll, R.J., and Kipnis, V. (2001) "Efficient regression
calibration for logistic regression in main study/internal validation study
designs with an imperfect reference instrument." \emph{Stat. Med.}
\strong{20}: 139-160.
}
