% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meuc.R
\docType{package}
\name{meuc}
\alias{meuc}
\alias{meuc-package}
\title{Fit Poolwise Regression Models}
\description{
Functions for implementing corrective methods for covariate measurement
error/missing data/unmeasured confounding. Methods include regression
calibration, propensity score calibration, and maximum likelihood based on
either two or three specified regression models.
}
\details{
\tabular{ll}{
Package: \tab meuc \cr
Type: \tab Package \cr
Version: \tab 1.1.1 \cr
Date: \tab 2018-06-02 \cr
License: \tab GPL-3 \cr
}
}
\references{
Acknowledgment: This material is based upon work supported by the National
Science Foundation Graduate Research Fellowship under Grant No. DGE-0940903.
}
\author{
Dane R. Van Domelen \cr \email{vandomed@gmail.com}
}
