% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psc.R
\name{psc}
\alias{psc}
\title{Propensity Score Calibration}
\usage{
psc(all_data = NULL, main = NULL, internal = NULL, external = NULL,
  y_var, x_var, gs_vars, ep_vars, tdm_family = "gaussian", surrogacy = TRUE,
  weighted = FALSE, ep_refit = FALSE)
}
\arguments{
\item{all_data}{Data frame with data for main study and validation study.}

\item{main}{Data frame with data for the main study.}

\item{internal}{Data frame with data for internal validation study.}

\item{external}{Data frame with data for the external validation study.}

\item{y_var}{Character string specifying name of Y variable.}

\item{x_var}{Character string specifying name of X variable.}

\item{gs_vars}{Character vector specifying names of variables for gold
standard propensity score.}

\item{ep_vars}{Character vector specifying names of variables for error-prone
propensity score.}

\item{tdm_family}{Character string specifying family of true disease model
(see \code{\link[stats]{glm}}).}

\item{surrogacy}{Logical value for whether to assume surrogacy, which means
that the error-prone propensity score is not informative of Y given X and the
gold standard propensity score. If validation data is external, have to
assume surrogacy.}
}
\description{
Implements propensity score calibration as described by Sturmer et al.
(\emph{Am. J. Epidemiol.} 2005). Requires validation data.
}
\details{
The disease model is a GLM:

g[E(Y)] = beta_0 + beta_x X + beta_g G

where G = P(X|\strong{C},\strong{Z}), with \strong{C} but not \strong{Z}
available in the main study.

In a validation study with (X, \strong{C}, \strong{Z}), logistic regression
is used to obtain fitted probabilities for G as well as an error-prone
version G* = P(X|\strong{C}). A linear model is fitted to map from G* to
E(G|G*). Finally, in the main study, G*'s are calculated, then E(G|G*),
and the disease model is fit for Y vs. (X, E(G|G*)).
}
\references{
Sturmer, T., Schneeweiss, S., Avorn, J. and Glynn, R.J. (2005) "Adjusting
effect estimates for unmeasured confounding with validation data using
propensity score calibration." \emph{Am. J. Epidemiol.} \strong{162}(3):
279-289.
}
